ORV2-761: Remove code smell
