+++ frontend/src/common/apiManager/hooks.ts
@@ -1,4 +1,4 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { addOrbcError } from "./errorsAPI";
 
 /**
